% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{calcEnrichmentScore}
\alias{calcEnrichmentScore}
\title{Calculate Enrichment Score}
\usage{
calcEnrichmentScore(geneList, geneSet)
}
\arguments{
\item{geneList}{a named vector containing the values of gene expression}

\item{geneSet}{a vector containing genes to represent a gene set}
}
\value{
a numeric indicating enrichment score
}
\description{
This function calculates enrichment score of a gene list on a specific gene set.
}
\examples{
geneList <- setNames(rnorm(10), paste0("G", seq_len(10)))
geneSet <- c("G1", "G5")
calcEnrichmentScore(geneList, geneSet)
}
