% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossValidataion}
\alias{crossValidataion}
\title{Cross Validation of DeepCC Model}
\usage{
crossValidataion(fs, labels, fold = 5)
}
\arguments{
\item{fs}{a data.frame containing functional spectra of patients (each row presents one sample)}

\item{labels}{a character vector containing training lables}

\item{fold}{a integer indicating the fold number of cross validation (defaut = 5)}
}
\value{
a numeric indicating error rate in a single run
}
\description{
This function performs cross validation of DeepCC Mode on the training data.
}
\examples{
crossValidataion(tcga.fs, tcga.labels)
}

