% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Transform scientific notation to expression form}
\usage{
fancy_scientific(l, digits = 3)
}
\description{
Transform scientific notation to expression form
}

