% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{get_DeepCC_features}
\alias{get_DeepCC_features}
\title{Get DeepCC Features}
\usage{
get_DeepCC_features(DeepCCModel, fs)
}
\arguments{
\item{DeepCCModel}{a trained DeepCC model}

\item{fs}{a data.frame containing functional spectra (each row presents one sample)}
}
\value{
a data.frame containing DeepCC Features extracted from the second-to-last layer
}
\description{
This function obtains DeepCC Features from functional spectra using the
second-to-last layer of the classifier in inference mode.
}
