% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{get_DeepCC_prob}
\alias{get_DeepCC_prob}
\title{Get DeepCC Prob Matrix}
\usage{
get_DeepCC_prob(DeepCCModel, newData)
}
\arguments{
\item{DeepCCModel}{a trained DeepCC model}

\item{newData}{a data.frame containing functional spectra of new data}
}
\value{
a matrix containing class probabilities for each sample
}
\description{
Get DeepCC Prob Matrix
}
