% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{getDeepCCProb}
\alias{getDeepCCProb}
\title{Get DeepCC prob mat}
\usage{
getDeepCCProb(DeepCCModel, newData)
}
\description{
Get DeepCC prob mat
}
\examples{
getDeepCCLabels(deepcc.model, newdata.fs)
}
