% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFigures.R
\name{plotKMCurve}
\alias{plotKMCurve}
\title{Plot Kaplan-Meier Curve}
\usage{
plotKMCurve(clinical, labels, annot = NULL, color = NULL, font = 10,
  xlab = "Follow up (weeks)", ylab = "DFS (prob.)")
}
\arguments{
\item{clinical}{a survival object created by Surv function in survival package}

\item{labels}{a vector containing subtyping labels of patients}

\item{annot}{a character indicating the annotation showed up in the plot}

\item{color}{a vector containing colors used for different subtypes}

\item{font}{a numeric indicating the font size used in the plot}

\item{xlab}{a character indicating the label of x-axis (default: "Follow up (weeks)")}

\item{ylab}{a character indicating the label of y-axis (default: "DFS (prob.)")}
}
\value{
a ggplot2 object of the plot
}
\description{
This function plots Kaplan-Meier Curve for survival analysis.
}
\examples{
clinical <- Surv(t.rfs, e.rfs)
color.cms <- c("#E69E00","#0070B0","#CA78A6", "#009C73")
plotKMCurve(clinical, labels, "GSE39582", color.cms)
}

