% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{preprocessGeneList}
\alias{preprocessGeneList}
\title{Preprocess Gene List}
\usage{
preprocessGeneList(geneList)
}
\arguments{
\item{geneList}{a named vector containing the value of gene expression}
}
\value{
a named vecter containing the value of gene expression
}
\description{
This function preprocess gene list for futhur process
}
\examples{
\dontrun{
geneList <- setNames(rnorm(10), paste0("G", seq_len(10)))
preprocessGeneList(geneList)
}
}
