% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{save_DeepCC_model}
\alias{save_DeepCC_model}
\title{Save DeepCC Model}
\usage{
save_DeepCC_model(deepcc_model, prefix)
}
\arguments{
\item{deepcc_model}{a trained DeepCC model}

\item{prefix}{file path prefix; outputs \code{prefix.hdf5} and \code{prefix.RData}}
}
\description{
Save DeepCC Model
}
