% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{trainDeepCCModel}
\alias{trainDeepCCModel}
\title{Train DeepCC Model}
\usage{
trainDeepCCModel(trainData, trainLabels, hiddenLayers = c(2000, 500, 120, 30,
  10), cores = NULL, device = "cpu")
}
\arguments{
\item{trainData}{a data.frame containing functional spectra of training data (each row presents one sample)}

\item{trainLabels}{a character vector containing lables of training data}

\item{hiddenLayers}{a numeric vector containing hidden layers architecture}

\item{cores}{a integer indicating how many GPUs used in parallel computing using GPUs (defaut = 1)}

\item{device}{a string indicating using CPU or GPU (defaut = "cpu")}
}
\value{
a trained DeepCC model
}
\description{
This function trains DeepCC Model on the training data.
}
\examples{
trainDeepCCModel(tcga.fs, tcga.labels)
}

