% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vis_samples}
\alias{vis_samples}
\title{Visualization of samples}
\usage{
vis_samples(data, labels, color, guide_fill = "legend")
}
\arguments{
\item{data}{data.frame}

\item{labels}{groups}

\item{color}{color}

\item{guide_fill}{legend}
}
\value{
a ggolot2 object
}
\description{
This function visualize samples.
}
\examples{
\dontrun{
set.seed(42)
df <- as.data.frame(matrix(rnorm(30*10), nrow=30, ncol=10))
labels <- sample(c("A", "B", "C"), 30, replace=TRUE)
color <- c(A="red", B="blue", C="green")
sample_plot <- vis_samples(df, labels, color)
}
}
