% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{get_DeepCC_label}
\alias{get_DeepCC_label}
\title{Get DeepCC Labels}
\usage{
get_DeepCC_label(DeepCCModel, newData, cutoff = 0.5, prob.mode = F,
  prob.raw = F)
}
\arguments{
\item{DeepCCModel}{a trained DeepCC model}

\item{newData}{a data.frame containing functional spectra of new data (each row presents one sample)}

\item{cutoff}{a numeric indicating cutoff of poster probability}
}
\value{
a character vector containing lables of training data
}
\description{
This function classifys new data set using trained DeepCC model.
}
\examples{
get_DeepCC_label(deepcc.model, newdata.fs)
}
