% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{get_DeepCC_prob}
\alias{get_DeepCC_prob}
\title{Get DeepCC prob mat}
\usage{
get_DeepCC_prob(DeepCCModel, newData)
}
\description{
Get DeepCC prob mat
}
\examples{
getDeepCCLabels(deepcc_model, newdata_fs)
}
