% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_sets}
\alias{get_gene_sets}
\title{Get Gene Sets}
\usage{
get_gene_sets(file)
}
\arguments{
\item{file}{filename of the gmt file}
}
\value{
a list containing gene sets by EntrezID
}
\description{
This function extract a list of gene sets from gmt file.
}
\examples{
msigdbv51 <- get_gene_sets("msigdb.v5.1.entrez.gmt")
}
