% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdbr.R
\name{get_msigdbr}
\alias{get_msigdbr}
\title{get MSigDBr from R package `msigdbr`}
\usage{
get_msigdbr(cores)
}
\arguments{
\item{cores}{a integer indicating cpu cores used in parallel computing (default = all cores -2 )}
}
\value{
a list containing 25, 724 gene sets, each sets contains multiple entrez_gene
}
\description{
This function defines MSigDBr from a R package, with 25, 724 gene sets
}
\examples{
MSigDBr <- get_msigdbr()
}
