% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{getDeepCCFeatures}
\alias{getDeepCCFeatures}
\title{Get DeepCC Features}
\usage{
getDeepCCFeatures(DeepCCModel, fs)
}
\arguments{
\item{DeepCCModel}{a trained h2o deep learning model}

\item{fs}{a data.frame containing functional spectra (each row presents one sample)}
}
\value{
a data.frame containing DeepCC Features extracted from the last hidden layer
}
\description{
This function obtains DeepCC Features from functional spectra.
}
\examples{
getDeepCCFeatures(deepcc.model, fs)
}

