% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectra}
\alias{getFunctionalSpectra}
\title{Generate Functional Spectra}
\usage{
getFunctionalSpectra(eps, geneSets = MSigDB, cores = parallel::detectCores()
  - 1)
}
\arguments{
\item{eps}{a data.frame containing gene expression profiles (each row presents one sample)}

\item{geneSets}{a List containg gene sets (defaut = MSigDB v5.0)}

\item{cores}{a integer indicating cpu cores used in parallel computing (defaut = all cores - 1)}
}
\value{
a data.frame containing functional spectra
}
\description{
This function generates functional spectra for given gene expression profiles.
}
\examples{
getFunctionalSpectra(eps)
}

