% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectra}
\alias{getFunctionalSpectra}
\title{Generate Functional Spectra}
\usage{
getFunctionalSpectra(eps, geneSets = "MSigDBv7", scale = T,
  cores = parallel::detectCores() * 0.8)
}
\arguments{
\item{eps}{a data.frame containing gene expression profiles (each row presents one sample)}

\item{geneSets}{a List containg gene sets (default: MSigDB v7.0)}

\item{cores}{a integer indicating cpu cores used in parallel computing (default = all cores - 1)}
}
\value{
a data.frame containing functional spectra
}
\description{
This function generates functional spectra for given gene expression profiles.
}
\examples{
getFunctionalSpectra(eps)
}
\seealso{
\code{\link{getFunctionalSpectrum}} for a single gene expression profile.
}
