% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectrum}
\alias{getFunctionalSpectrum}
\title{Generate Functional Spectrum}
\usage{
getFunctionalSpectrum(
  expressionProfile,
  geneSets = "MSigDBv7",
  refExp = NULL,
  logChange = FALSE,
  inverseRescale = FALSE,
  filter = -3
)
}
\arguments{
\item{expressionProfile}{a named numeric vector containing gene expression profile}

\item{geneSets}{a List containing gene sets (default: MSigDB v7)}

\item{refExp}{a character indicating cancer typer according to TCGA's indentifier, or a named vector reference expression}

\item{logChange}{a logical flag indicating whether the input data is already in log change form, e.g., for two color microarray, you should turn it on. (default: FALSE)}

\item{inverseRescale}{a logical flag indicating whether we rescale the reference to the scale of input data. If your single sample is microarray data and the reference is RNA-Seq, you should turn it on. (default: FALSE)}

\item{filter}{a numeric indicating the cutoff value of expression. (default: -3)}
}
\value{
a numeric vector containing functional spectrum
}
\description{
This function generates functional spectrum for a single gene expression profile.
}
\note{
You can generate the reference expression profile from your previous data or public data, which is the same(similiar) cancer type and platform.
In DeepCC we also prepared average expression profiles of each cancer types in TCGA project as references. To use them, just use the TCGA identifier (COADREAD, BRCA, OV, etc.) to indicate the cancer type.
If your single sample is microarray data, we strongly sugguest turn the parameter \code{inverseRescale} on, since TCGA is RNA-Seq, which has very small expression value for low expressed genes, compared with microarray.
}
\examples{
\dontrun{
ep <- setNames(rnorm(100), paste0("G", seq_len(100)))
ref <- setNames(rnorm(100), paste0("G", seq_len(100)))
fs <- getFunctionalSpectrum(ep, geneSets=list(setA=c("G1","G5","G20")), refExp=ref)
}
}
\seealso{
\code{\link{getFunctionalSpectra}} for a batch of gene expression profiles.
}
