% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectrum}
\alias{getFunctionalSpectrum}
\title{Generate Functional Spectrum}
\usage{
getFunctionalSpectrum(expressionProfile, geneSets = MSigDB, refExp,
  logChange = F)
}
\arguments{
\item{expressionProfile}{a named numeric vector containing gene expression profile}

\item{geneSets}{a List containg gene sets (defaut = MSigDB v5.0)}

\item{refExp}{a character indicating cancer typer according to TCGA's indentifier, or a named vector containing reference expression}

\item{cores}{a integer indicating cpu cores used in parallel computing (defaut = all cores - 1)}
}
\value{
a numeric vector containing functional spectrum
}
\description{
This function generates functional spectrum for a single given gene expression profile.
}
\examples{
getFunctionalSpectrum(ep)
}

