% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{getFunctionalSpectrum}
\alias{getFunctionalSpectrum}
\title{Generate Functional Spectrum}
\usage{
getFunctionalSpectrum(expressionProfile, geneSets = MSigDB, refExp,
  logChange = F, inverseRescale = F, filter = -3)
}
\arguments{
\item{expressionProfile}{a named numeric vector containing gene expression profile}

\item{geneSets}{a List containg gene sets (defaut = MSigDB v5.0)}

\item{refExp}{a character indicating cancer typer according to TCGA's indentifier, or a named vector containing reference expression}

\item{logChange}{a logical flag indicating whether the input data is already in log change form, e.g., for two color microarray, you should turn it on. (default: FALSE)}

\item{inverseRescale}{a logical flag indicating whether we rescale the reference to the scale of input data. If your single sample is microarray data and the reference is RNA-Seq, you should turn it on. (default: FALSE)}

\item{filter}{a numeric indicating the cutoff value of expression. (default: -3)}
}
\value{
a numeric vector containing functional spectrum
}
\description{
This function generates functional spectrum for a single given gene expression profile.
}
\examples{
getFunctionalSpectrum(ep)
}

