% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalSpectra.R
\name{preprocessGeneList}
\alias{preprocessGeneList}
\title{Preprocess Gene List}
\usage{
preprocessGeneList(geneList)
}
\arguments{
\item{geneList}{a named vector containing the value of gene expression}
}
\value{
a named vector containing the value of gene expression
}
\description{
This function preprocess gene list for futhur process.
}
\examples{
preprocessGeneList(geneList)
}
