% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{train_DeepCC_model}
\alias{train_DeepCC_model}
\title{Train DeepCC Model}
\usage{
train_DeepCC_model(
  trainData,
  trainLabels,
  epochs = 60,
  dropout = 0.4,
  activation_func = "selu",
  validation_split = 0.2
)
}
\arguments{
\item{trainData}{a data.frame containing functional spectra of training data (each row presents one sample)}

\item{trainLabels}{a character vector containing lables of training data}

\item{epochs}{the number of epochs}

\item{dropout}{dropout rate}

\item{activation_func}{activation funtion}
}
\value{
a trained DeepCC model
}
\description{
This function trains DeepCC Model on the training data.
}
\examples{
train_DeepCC_model(tcga_fs, tcga_labels)
}
