% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLearning.R
\name{trainDeepCCModel}
\alias{trainDeepCCModel}
\title{Train DeepCC Model}
\usage{
trainDeepCCModel(trainData, trainLabels, hiddenLayers = c(2000, 500, 120, 30,
  10), cores = parallel::detectCores() - 1)
}
\arguments{
\item{trainData}{a data.frame containing functional spectra of training data (each row presents one sample)}

\item{trainLabels}{a character vector containing lables of training data}

\item{hiddenLayers}{a numeric vector containing hidden layers architecture}

\item{cores}{a integer indicating cpu cores used in parallel computing (defaut = all cores - 1)}
}
\value{
a trained h2o deep learning model
}
\description{
This function trains DeepCC Model on the training data.
}
\examples{
trainDeepCCModel(tcga.fs, tcga.labels)
}

