% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vis_samples}
\alias{vis_samples}
\title{Visualization of samples}
\usage{
vis_samples(data, labels, color, guide_fill = "legend")
}
\arguments{
\item{data}{data.frame}

\item{labels}{groups}

\item{color}{color}

\item{guide_fill}{legend}
}
\value{
a ggolot2 object
}
\description{
This function visualize samples.
}
\examples{
sample_plot <- vis_samples(df, labels, color)
}
