% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vis_samples}
\alias{vis_samples}
\title{Visualization of samples}
\usage{
vis_samples(data, labels, color, guide_fill = "legend")
}
\value{
a list containing gene sets by EntrezID
}
\description{
This function visualize samples.
}
